Remove height from view content
With the new fixed view container, there was missing padding at the bottom of view content.

I solved it by adding an empty div with a height of 1rem beneath view content, but that seemed messy. Removing height property here was a simpler solution.

I'm not sure why this style was added, but removing it had no discernable impact beyond solving the issue above.